+++ pymc_marketing/__init__.py
@@ -27,7 +27,7 @@ import pymc_marketing.data.fivetran  # noqa: E402
 
 # Register SpecialPrior deserializers (LogNormalPrior, LaplacePrior, etc.)
 import pymc_marketing.special_priors  # noqa: E402, F401
-from pymc_marketing import bass, clv, customer_choice, mmm  # noqa: E402
+from pymc_marketing import bass, clv, customer_choice, mmm, pie  # noqa: E402
 from pymc_marketing.version import __version__  # noqa: E402
 
-__all__ = ["__version__", "bass", "clv", "customer_choice", "mmm"]
+__all__ = ["__version__", "bass", "clv", "customer_choice", "mmm", "pie"]
